test: PROJECTTEMPLATE-69 Рефакторинг модуля для тестов
